java italian iter_1

+++ src/main/java/org/636767031a6d9265ec0178e6/636767031a6d9265ec0178e6.java
@@ -15,7 +15,7 @@ public class ByteArrayConverter {
         }
         byte[] result = new byte[array.length];
         for (int i = 0; i < array.length; i++) {
-            result[i] = Objects.requireNonNull(array[i], "Element at index " + i + " is null").byteValue();
+            result[i] = Objects.requireNonNull(array[i], "Element at index " + i + " is null");
         }
         return result;
     }
